Add success message when request submitted
