+++ src/pages/OfficeHours/RequestDeskCreditPage/index.ts
@@ -1,0 +1,1 @@
+export { RequestDeskCreditPage } from "./RequestDeskCreditPage";
